Initial Tauri Rework
This is the basis of the re-write as I transition this app from the custom backend to a Tauri application. It has most of the primary functionality of the application, but redux is still used as I haven't entirely phased it out. Some important things such as the stdout console and joysticks are not yet implemented. This is still a WIP and more is to come